requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.103.0
mangum==0.15.0  # Include this if deploying on AWS Lambda